clase 4/2/2026

+++ AppEmpresa/departamentos/borrar.php
@@ -3,11 +3,13 @@ session_start();
 require_once "../includes/auth.php";
 require_once "../includes/db.php";
 require_once "../includes/functions.php";
+require_login();
+require_admin();
 
 $id = intval($_GET['id'] ?? 0);
 if ($id) {
     $stmt = $bd->prepare("DELETE FROM departamentos WHERE CodDept = ?")->execute([$id]);
     flash_set("Departamento borrado");
 }
-header("Location: listar.php");
-exit;
+redirect("listar.php");
+
